first implementation of sceneManager.
in this implementation, I had first repeated the whole system and entity manager, together with the components array. this basically would create a mini-engine for each scene. this of course causes a big overhead since the engine was not done thinking of this archtecture

but as a proof of concept, it worked